requirements.txt
pandas
pandas-ta
ccxt
requests
aiogram
Pillow
aiohttp
asyncio
boto